GreatHistory/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Faker==9.8.0
idna==3.3
Markdown==3.3.4
mixer==7.2.0
Pillow==8.4.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7
